// packages/image-generation/src/types.ts
export interface ImageInputFile {
  data: Buffer;
  name: string;
  type: string;
  url?: string;
  storageBucket?: string;
  storageKey?: string;
  imageFileId?: string;
}